eval/requirements.txt
Flask
python-dateutil
easydict
scikit-image
azure-storage-blob
applicationinsights
libmc
waitress